requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.55.0
    # via matplotlib
iniconfig==2.0.0
    # via pytest
kiwisolver==1.4.7
    # via matplotlib
matplotlib==3.9.2
    # via -r requirements.in
numpy==2.1.3
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   numpy-stl
    #   scipy
numpy-stl==3.1.2
    # via -r requirements.in
packaging==24.2
    # via
    #   matplotlib
    #   pytest
pillow==11.0.0
    # via
    #   -r requirements.in
    #   matplotlib
pluggy==1.5.0
    # via pytest
pyparsing==3.2.0
    # via matplotlib
pytest==8.3.3
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via matplotlib
python-utils==3.9.0
    # via numpy-stl
scipy==1.14.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
typing-extensions==4.12.2
    # via python-utils
